created basic types for Support Vector Machines
Right now, this is all the stuff I can say for certain should be in an SVM implementation. I have a way to specify one's own kernel, a way to define one's own samples, and a Classifier which the SVM solver should generate.

I am not a master at numerical analysis, so the next step will probably be ugly. I need a way to solve for a Classifier given a Problem. For now, I will probably try a hacky approach.