refined SolveRandomly for better clarity
Now, SolveRandomly is clearer (mathematically) and has been renamed to better reflect its purpose. Now it actually performs worse than it did before, because I assumed before that the solution would be a linear combination of the sample vectors. Now, I make no such assumption because I want it to be as mathematically simple as possible.