renamed CoordDescentSolver to GradientDescentSolver
I believed that my algorithm was called "coordinate descent", but that would mean only following the gradient for one coordinate at a time. This algorithm is pretty much an old-school gradient descent algorithm, with the addition that it projects the input vector onto the constraints after every iteration. Thus, it has some attributes of a coordinate descent solver, but more attributes of a gradient descent solver.