I suck at calculus
Before now, I had taken the derivative of the Langrangian Dual incorrectly. As a result, I was always getting twice the canonical answer.

When you have a product like (c1 + c2 + c3 + ...)^2, I forgot that c1*c2 will appear twice, once in each order. Now I account for that.

Interestingly, I do not need to take special account of the product c1*c1. This is because the power rule says that d/dc1 of (c1*c1) is 2*c1. My existing code for computing the derivative, which doesn't even notice when it's multiplying the constant by itself, already adds this exact term to the partial.