perform SVM caching in Kernel
In order to implement a Kernel which caches inner products a way to look up a Sample in a map. Before this change, Samples were simply []float64's, and Go has no way to compare two of these by pointer. To deal with this, I created a way to tag Samples with a UserInfo field. This way, a simple map lookup can be used using the UserInfo. It also allows the user to mark vectors that they do not want to cache; such vectors have UserInfo == 0.