don't walk along near-zero gradients in SVM solver
When the gradient is pretty much only a linear combination of the active constraints, projecting out the constraints *should* result in a gradient of zero. Of course, floating points never let us off easy. Instead, we get a projected gradient in a semi-random direction. Computing constraints with such a gradient isn't reliable, since we can't even guarantee that it will be optimal to stop in the direction of the gradient (as opposed to going backwards), and the gradient is likely not to be perfectly orthogonal to our constraints.