improve expectation computation performance
Profiling showed that 80% of training was typically spent in Matrix.Mul(). I quickly set out to optimize these matrix multiplications.

Unoptimized matrix multiplication for ExpectedVisible/ExpectedHidden was huge overkill, since we were multiplying a vector of booleans (0 or 1) by a matrix. I was able to utilize the fact that input vectors are binary vectors to speed up matrix multiplication by completely doing away with floating-point multiplication.

As a result of this optimization, training seems to run about 2x faster than before (according to the new benchmark).