remove overhead for SGD with batch size = 1
When GOMAXPROCS was greater than 1 but BatchSize was 1, there was a lot of overhead (at least in demos like old_demos on OS X) for channel sending/receiving. This was only a problem when the samples are small-ish and can get through a neuralnet very quickly (quickly enough that a few 100ns for sending/receiving matters), but it was a problem nonetheless.

To fix the problem I just described, the SGD trainer uses a special no-currency routine if the BatchSize is 1 (or 0). This ensures that old code isn't made slower by the addition of concurrent training.