simplified RNN Gradient interface
Simple motivation: the gradient is a glorified vector, so why not treat it as such? It should be as easy to implement Gradient as possible.

Future-oriented motivation: for techniques like RMSProp that look at the gradient component-wise, it is necessary to be able to access the components of a Gradient. The most efficient way to do this is through straightup slices.