fixed RNN training bugs
This fixes two bugs in the SGD trainer. First, the trainer did not reset the networks after back propagation, which used to be fine because StepGradient() called Reset(), but is no longer fine with mini-batches. Also, the concurrent trainer had a bug where it might first call the synchronous trainer and then also train concurrently. This caused nil-reference problems if s.BatchSize == 0.