change the way Batcher and SGD() work
The new Batcher is optimized for batches sizes significantly larger than GOMAXPROCS (e.g. 100). This is fairly common in practice, so the optimizations make sense. It runs about twice as fast (single threaded) as it did before, and even faster multithreaded.

I am still seeing that, for neural nets with many parameters, adding parallelism doesn't always increase performance. I suspect that this is a memory issue: there are too many parameters to fit in cache, so there is a lot of contention for memory, and memory across the system is becoming the bottleneck. I will produce some graphs to pinpoint the memory bottleneck more exactly.