attempted to make stacked_block.go easier to read
This is a really confusing API to implement because it involves packing state vectors together in bigger state vectors, but then these aggregate state vectors are themselves in different "lanes" for the batch. States need to be split, rejoined, added to gradient maps, etc., and this is all pretty mind boggling.

To help make it easier to understand what a StackedBlock is doing, I did a few things:

 * Use more verbose variable names than normal
 * Put tons of functionality into sub-routines
 * Carefully select hidden function names
 * Use more doc comments than usual.