swap RNN sequence sort in FullRGradienter
This may increase performance in some pathological scenarios. The idea is to get the longest sequences out of the way first, so one or two Goroutines get stuck with them while the rest of the Goroutines can work on all the smaller sub-batches of sequences. If the longest sequence was left to the end, all the other Goroutines would have no work to do as one Goroutine worked on the long sequence.