flipped cross entropy cost for pure 0/1 samples
Before, CrossEntropyCost would always give NaN when the expected sample contained a 0 (pretty much something ubiquitous in ML). Now, it only gives NaN when the actual sample from the previous layer is exactly 0 or 1 (which should not *completely* happen if the previous layer is a sigmoid or softmax, although it may still be a risk that's worth addressing in the future).