make SampleSet an interface
There is a very specific reason I did this. Sometimes, a sample might just represent a piece of text or something. When passing this piece of text to a neural net, it needs to be converted into a huge vector (possibly much larger than the original text). If SampleSet was a simple slice, this vector conversion would need to be done in advance, and the vector would need to be stored in memory. This is impractical for many applications. By making SampleSet an interface, the GetSample() method can dynamically convert the compact, easy-to-store sample into a vectorized value.