feed LSTM input into next input gates
Before now I had misunderstood LSTMs. I was feeding the raw, unsquashed and ungated output state back into the inputs of the next time step (along with the time step's actual inputs). Now I realize that, in traditional LSTMs, the output of the LSTM is also used as part of the input for the next round of gates, since the internal state would be unsquashed and might quickly saturate the gate neurons.