activation functions now implement autofunc.RBatcher
This is mostly a memory optimization for gated RNNs. When batching a gated RNN before now, all inputs to sigmoids were sliced and then the outputs reconcatenated. This caused a lot of extra memory consumption, which is especially bad when dealing with batches of long sequences (since sequence memory is the main source of memory pressure).