fixed potential hanging bug when training big trees
The bug was due to rounding errors when adding up entropy. It is not numerically sound to rely solely on the entropy, since no change in entropy could appear to be a miniscule improvement in entropy due to rounding errors when adding up sub-entropies. By checking if the best split doesn't do anything productive, we can more robustly terminate the algorithm.