made "seqtoseq" sub-package of "rnn" package
This is a major organizational change which will break pretty much all code which depends on the RNN package. However, I believe this change is worth it because it removes the direct link between RNNs and sequence-to-sequence training. When I made the original RNN package, I didn't realize how many ways there were to train RNNs besides in a seq-to-seq style.

I think the pattern I will use going forward is that RNNs themselves will be implemented in the rnn package, but training algorithms (e.g. CTC, seq-to-seq) will have their own packages.