learnable initial states for RNNs
This is a major, breaking change. However, it is necessary to make my recurrent models as powerful as those used in the literature.

Most recurrent models have the notion of a "state", and, before now, I always initialized this state to 0. Some models can deal with this 0-initialization very well; powerful models can learn to make a 0-state have a very specific meaning. However, it may be benificial to pretty much any model to learn the initial state. This way, the meanings of each state value can be defined and learned naturally, without being constrained to the notion that "0 means nothing in the past has been seen".

This change adds methods to the Block interface, and said interface is very common and used heavily throughout this project and descendant projects. Consequently, this change will break any external package which defines a Block, or any package which manually uses a block and initializes its state to 0.

This change also adds more internal fields to LSTM blocks, network blocks, and GRU blocks. This means that the serialization logic is different. As a consequence, all previously serialized recurrent neural networks will no longer be loadable.