less memory overhead for ConvLayer
Before, the ConvLayer's Result stored the matrixified (redundant) input tensor. In pretty much any conv net, the result of turning the input tensor into a matrix is a few times larger than the original tensor's backing data (due to the redundancy). As a result, doing a forward+backward pass through a deep convolutional network used a lot of memory.

By re-computing the matrix representation of the input during the backward pass, the network uses less memory at a time. This is important for large networks with a large batch size.