drafted new LSTM implementation
The old LSTM implementation kind of a mess--difficult to understand, full of manually pooling, etc. The new implementation is a bit longer, but it is much simpler to read (at least, in my opinion). There is less vector voodoo, partly thanks to the fact that we can use custom state objects, and partly thanks to my willingness to add helper functions to autofunc.