provide PropagateStart with start states
This is necessary in one important scenario. A Block may wish to store some internal state in a variable and then squash this state before using it (e.g. turn an unrestricted variable into a vector of probabilities). Then the StartState() or StartRState() methods need to do some computation, and it would be costly to re-do that computation in order to back-propagate through it in PropagateStart.

The bigger problem is that PropagateStartR() needs to pass a Block like the one described above the original RVector, or else the block won't know how to back-propagate through the squashed state (the result depends on the unrestricted variable's r-gradient). However, passing the RVector directly would be silly, since no other PropagateGradient routines take RVectors. Instead, it can pass the initial RState to the PropagateStartR() method, and the RState should keep track of the RVector as appropriate.