use diagonal matrix for peephole weights
This is what everyone seems to do with peepholes. I guess peepholes are mostly useful for counting types of things, and those applications only require access to the current cell. This change should reduce the number of parameters in an LSTM by half, while keeping most of the functionality (and probably greatly improving training performance).